fix: bugfix for schedule time
