+++ src/components/Compose/SchedulePostSettings.tsx
@@ -10,6 +10,7 @@ import { ClickableButton } from '@/components/ClickableButton.js';
 import { DatePicker } from '@/components/DatePicker.js';
 import { TimePicker } from '@/components/TimePicker.js';
 import { queryClient } from '@/configs/queryClient.js';
+import { checkScheduleTime, CreateScheduleError } from '@/helpers/checkScheduleTime.js';
 import { enqueueErrorMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
 import { fireflySessionHolder } from '@/providers/firefly/SessionHolder.js';
@@ -40,6 +41,7 @@ export const SchedulePostSettings = memo<SchedulePostSettingsProps>(function Sch
             }
 
             if (task) {
+                checkScheduleTime(value);
                 const result = await FireflySocialMediaProvider.updateScheduledPost(task.uuid, value);
                 if (!result) return;
                 queryClient.refetchQueries({
@@ -51,9 +53,13 @@ export const SchedulePostSettings = memo<SchedulePostSettingsProps>(function Sch
 
             onClose();
         } catch (error) {
-            enqueueErrorMessage(getSnackbarMessageFromError(error, t`Failed to set schedule time.`), {
-                error,
-            });
+            if (error instanceof CreateScheduleError) {
+                enqueueErrorMessage(error.message);
+            } else {
+                enqueueErrorMessage(getSnackbarMessageFromError(error, t`Failed to set schedule time.`), {
+                    error,
+                });
+            }
             throw error;
         }
     }, [value, task, onClose]);
