+++ src/services/crossSchedulePostThread.ts
@@ -1,17 +1,17 @@
 import { t } from '@lingui/macro';
 import { delay } from '@masknet/kit';
-import dayjs from 'dayjs';
-import { compact, first } from 'lodash-es';
+import { first } from 'lodash-es';
 
 import type { SocialSourceInURL } from '@/constants/enum.js';
-import { readChars } from '@/helpers/chars.js';
+import { checkScheduleTime, CreateScheduleError } from '@/helpers/checkScheduleTime.js';
 import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getProfileSessionsAll } from '@/helpers/getProfileState.js';
+import { getScheduleTaskContent } from '@/helpers/getScheduleTaskContent.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
 import type { SchedulePayload } from '@/helpers/resolveCreateSchedulePostPayload.js';
 import { fireflySessionHolder } from '@/providers/firefly/SessionHolder.js';
 import { FireflySocialMediaProvider } from '@/providers/firefly/SocialMedia.js';
-import { CreateScheduleError, createSchedulePostsPayload } from '@/services/crossSchedulePost.js';
+import { createSchedulePostsPayload } from '@/services/crossSchedulePost.js';
 import { uploadSessions } from '@/services/metrics.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
 
@@ -20,11 +20,7 @@ export async function crossPostScheduleThread(scheduleTime: Date) {
         const { posts, type } = useComposeStateStore.getState();
         if (posts.length === 1) throw new Error(t`A thread must have at least two posts.`);
 
-        if (dayjs().add(7, 'day').isBefore(scheduleTime)) {
-            throw new CreateScheduleError(t`Up to 7 days can be set as the scheduled time.`);
-        } else if (dayjs().isAfter(scheduleTime, 'minute')) {
-            throw new CreateScheduleError(`The scheduled time has passed.`);
-        }
+        checkScheduleTime(scheduleTime);
 
         const results = new Map<
             SocialSourceInURL,
@@ -46,14 +42,7 @@ export async function crossPostScheduleThread(scheduleTime: Date) {
         const postsPayload = [...results.values()];
 
         const post = first(posts);
-        const assets = compact([
-            post?.images.length ? t`[Photo]` : undefined,
-            post?.video ? t`[Video]` : undefined,
-            post?.poll ? t`[Poll]` : undefined,
-        ]).join('');
-
-        const chars = post ? readChars(post.chars, 'visible') : '';
-        const content = `${chars}${chars.length > 0 ? '\n' : ''}${assets}`;
+        const content = getScheduleTaskContent(post);
 
         await uploadSessions(fireflySessionHolder.sessionRequired, getProfileSessionsAll());
